Lighter processing of skipped measurements
The functions `tag_values/1` and `measurement/{1,2}`, if set, are
executed each time a metric is incoming, regardless if the metric would
be actually reported or skipped.

This is a bit of a waste of processing power that may become
problematic in some scenarios. For example, if we collect `Oban` metrics,
but we are only interested in measuring only the results of a certain worker.
As it is, those functions would be invoked for every single oban job executed
in the system.

This PR solves that postponing the execution of such functions to after
`keep?` is evaluated to `true` and `measurement` is verified not `nil`.